// api/web-sdk/models/SmartObject.ts
export interface SmartObject {
    /**
     * Unique identifier for the smart object
     * @type {string}
     * @memberof SmartObject
     */
    id: string;
    /**
     * Organization identifier that owns this smart object
     * @type {string}
     * @memberof SmartObject
     */
    orgId: string;
    /**
     * Reference to the chain information
     * @type {Chain}
     * @memberof SmartObject
     */
    chain?: Chain;
    /**
     * System properties of the object
     * @type {object}
     * @memberof SmartObject
     */
    system?: object;
    /**
     * Custom properties of the object
     * @type {object}
     * @memberof SmartObject
     */
    custom?: object;
    /**
     * Custom hash for the object
     * @type {string}
     * @memberof SmartObject
     */
    customHash?: string;
    /**
     * Metadata information for the object
     * @type {Metadata}
     * @memberof SmartObject
     */
    metadata: Metadata;
    /**
     * List of assets associated with the object
     * @type {Array<Asset>}
     * @memberof SmartObject
     */
    assets?: Array<Asset>;
    /**
     * ID of the owner of the object
     * @type {string}
     * @memberof SmartObject
     */
    owner: string;
    /**
     * Template identifier that defines the structure and behavior of this smart object
     * @type {string}
     * @memberof SmartObject
     */
    templateId: string;
    /**
     * Location information for the object
     * @type {Location}
     * @memberof SmartObject
     */
    location?: Location;
    /**
     * Nonce value for the object
     * @type {number}
     * @memberof SmartObject
     */
    nonce: number;
    /**
     * Version of the object
     * @type {number}
     * @memberof SmartObject
     */
    version: number;
    /**
     * state root of the object
     * @type {string}
     * @memberof SmartObject
     */
    stateHash: string;
    /**
     * object content hash - metadata, assets, location, custom
     * @type {string}
     * @memberof SmartObject
     */
    contentHash: string;
    /**
     * Hash of the integrity data
     * @type {string}
     * @memberof SmartObject
     */
    integrityHash: string;
    /**
     * Hash of the previous integrity hash
     * @type {string}
     * @memberof SmartObject
     */
    prevIntegrityHash: string;
    /**
     * Timestamp of when the object was created
     * @type {Date}
     * @memberof SmartObject
     */
    whenCreated: Date;
    /**
     * Timestamp of the last modification
     * @type {Date}
     * @memberof SmartObject
     */
    whenModified: Date;
}